requirements.txt
absl-py==2.3.0
accelerate==1.8.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==1.4.3
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anthropic==0.55.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
bitsandbytes==0.45.5
bleach==6.2.0
blis==1.2.1
braceexpand==0.1.7
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.34.0
distro==1.9.0
e2cnn==0.2.3
einops==0.6.1
einops-exts==0.0.4
et_xmlfile==2.0.0
ever_beta==0.5.1
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.13
fastapi==0.115.14
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.14.0
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
gdcm==1.1
git-lfs==1.6
gitdb==4.0.12
GitPython==3.1.44
gradio==5.35.0
gradio_client==1.10.4
greenlet==3.2.3
groovy==0.1.2
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-socks==0.10.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
ipykernel==6.29.5
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
langchain==0.3.26
langchain-anthropic==0.3.16
langchain-community==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langcodes==3.5.0
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.72
langsmith==0.4.4
language_data==1.3.0
latex2mathml==3.78.0
lazy_loader==0.4
lightning-utilities==0.15.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mcp==1.12.0
mdurl==0.1.2
mistune==3.1.3
mmcv-full==1.7.1
-e git+https://github.com/waterdisappear/SARATR-X.git@8357897b02b30614e95cc3f5adf6779260fd50a4#egg=mmdet&subdirectory=detection
mmsegmentation==0.17.0
model-index==0.1.11
mpmath==1.3.0
multidict==6.6.2
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.45.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.93.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.18
ormsgpack==1.10.0
oss2==2.17.0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
peft==0.7.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.8
portalocker==3.2.0
preshed==3.0.10
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydicom==3.0.1
pydub==0.25.1
Pygments==2.19.2
pylibjpeg==2.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-socks==2.7.1
pytorch-ignite==0.5.2
pytz==2023.4
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.28.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.25.1
ruff==0.12.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.7.0
scipy==1.11.3
segmentation_models_pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@22d84a7180f299da877490cf79104f57b343f1af
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==5.0.0
sentencepiece==0.1.99
sentry-sdk==2.32.0
setproctitle==1.3.6
shapely==2.1.1
shellingham==1.5.4
shortuuid==1.0.13
simsimd==6.4.9
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
-e git+https://github.com/Ireliya/spatialreasonagent.git@15db70ced32c0f423ae6484af86f41520f7a63ea#egg=spatialReasonAgent
SQLAlchemy==2.0.41
srsly==2.5.1
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.5
svgwrite==1.4.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
terminaltables==3.1.10
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==0.9.12
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.12.0
torch==2.7.1
torchmetrics==1.8.0
torchvision==0.22.1
torchxrayvision==1.3.5
tornado==6.5.1
tqdm==4.65.2
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@88d960937c81a32bfb63356a2e8ecf7999619681
triton==3.3.1
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.35.0
wandb==0.20.1
wasabi==1.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xformers==0.0.31
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
